Beginning of per-player actors.
They are not finished and should not be added in this state, they work
but important questions & features are stll missing.
I didn't had the time & motiviation to finish this, as the codebase is a
"nightmare" to work with IMO.

I added all needed hints for someone to finish this.